+++ crates/modular-decomposition/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ### Added
 
++ Tests for `make_index` macro
 + `MDTree::nodes(module)` function
 
 ### Changed